refactor: Refactor the html and css code
Refactor the html code to make use of some bootstrap classes like
text-uppercase, also added rows and cols to every element in the mockup.
Refactor the css code to make use of more general classes like
text-muted and text-thin to achieve the same results as before, but in a
cleaner way.

Resolves #2